extract Semaphore to utils package
